requirements.txt
requests
google.cloud
google.cloud.vision
google.cloud.storage
boto3
boto